requirements.txt
python-dotenv
selenium
pymysql
chromedriver_autoinstaller
tk
tkinter-tooltip
requests
cryptography